requirements.txt
streamlit==1.33.0
openai>=1.0.0
pillow==10.2.0
requests==2.31.0
